+++ .github/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### New features since last release
 
+* The Jet interpreter for XIR scripts now accepts a `dimension` option for CV circuits. [(#47)](https://github.com/XanaduAI/jet/pull/47)
+
 * The Jet interpreter for XIR programs now supports a `probabilities` output. [(#44)](https://github.com/XanaduAI/jet/pull/44)
 
 * The `Circuit` class now supports taking the expected value of an observable. [(#46)](https://github.com/XanaduAI/jet/pull/46)